clipart_ops/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clipart-ops"
version = "0.1.0"
description = "Ứng dụng desktop Python quản lý bundle clipart Etsy, Google Drive và Pinterest."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "PySide6>=6.8.0",
  "pyqtdarktheme>=0.1.7",
  "pydantic>=2.10.0",
  "httpx>=0.28.0",
  "markdown>=3.7",
  "watchdog>=6.0.0",
  "google-api-python-client>=2.164.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
